requirements.txt
earthengine-api>=1.7
geeagri>1.0.0
geemap
importlib
matplotlib
numpy
optuna
pandas
pcse
pedon
PyYAML
SALib
scipy
seaborn
